Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

looks good to me (except style nit)

I am strongly convinced we should stay on the safe side and make this safety check also for non-interactive rebase. I don't want any user to unnecessarily
loose any of his work because of a skipped check in egit. 

Maybe advanced users like you could use it but I think it's not really necessary to accept this risk if we enhance this proposal with the following additional improvements:
- add an option to automatically stash / pop stash before/after rebase
- add another option to automatically use this option without asking the user

I think from the user's point of view this should be equivalent to the old implementation without risking data loss in case the user aborts the rebase.

Also, if we don't allow rebase with a dirty tree, we can be sure that users which use both EGit and native git on the same repository don't confuse native git with a situation it doesn't expect.

We should also enhance (in another change) the dialog to show the list of dirty file paths in a compact tree representation like we do in the staging view

Patch-set: 2
Label: Code-Review=-1