Update patch set 1
Patch Set 1:

> the cursor is flickering a lot so it seems there are tons of UI refreshes which we should probably suppress until the data is available

I don't know SWT well enough to know what we can do about that.

> initially after restarting Eclipse the collapsed tree of all staged files seems to be always fully expanded, is this intended ?

The default state is expanded and we don't remember the expansion state. Maybe we should store it in some kind of settings.

> Do you see similar behavior on your platform ?

Yes, I also didn't feel any difference.

I suspect the bottleneck is currently in getting the right IResource, see point 2 in the bug report. The code in ResourceUtil uses findFilesForLocationURI and co which I've heard is slow.

I'll work on that next. In hindsight, I should probably have started with that one :).

> Did you measure the performance difference ?

No, as these tree content provider methods are called from the UI when creating the tree, it's hard to measure before/after.

Apart from that, I feel that this version is nicer, e.g. look at how setParent was done before. Now we construct the tree once and then only read it.

Patch-set: 1
Label: Verified=0