Update patch set 2
Patch Set 2: Code-Review+1

+1 also for me for the idea of an automated stash/pop.

But one comment: That we lost some unsaved changes during a rebase abort is a severe bug which we have to track anyhow. I think we did some wrong checkout's after a rebase abort and I will hunt for this. It is not intrinsic to a a rebase abort that you'll loose unsaved changes if you have started with a dirty state.

Patch-set: 2
Reviewer: Gerrit User 5 <5@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1