Create Branch: Change source selection from combo to selection dialog
When a repository has many branches or tags, the selection with the
combo is ineffective (e.g. no filtering). Apart from that, it slows down
the dialog even when the user does not intend to change the source
branch.

This changes it to a "Select..." button which shows a branch selection
dialog where the user can filter.

To clean up the appearance, an icon is shown for the source branch
representing the type and the short branch/tag name is used (that's the
name the user sees everywhere else).

Bug: 422661
Change-Id: Icfaabed13f890f5b8abb0833e847a76e648f0b86
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>