[stagingView] Improve performance of switching presentation
On a test repository with 4000 changed files, switching between
presentations was 6 seconds before, 1 second after this change.

One problem was that expanding the previously expanded elements was
unnecessarily done when switching presentations.

Also, the refresh was done twice because run() is also called for the
deselected action. See bug 422910 about improving the documentation of
IAction.AS_RADIO_BUTTON.

Bug: 420825
Change-Id: Ie27f15b7a6eb4146f874312a3a6431cf52d5d4b2
Signed-off-by: Robin Stocker <robin@nibor.org>