Update patch set 1
Patch Set 1:

In the repositories view, select the 3.2.xxx tag of the jgit repo, and then do "Push Tag...". The tag should be checked and it should also be visible (tree scrolled to the end).

With just reveal as before, it doesn't work here on Linux, the tree's scroll position is at the start. With this change, it works.

Maybe it's a platform-specific problem?

Patch-set: 1
Label: Verified=0