Update patch set 1
Patch Set 1:

Hm, I didn't notice that. When trying to reproduce this, it only happened the first time the wizard was used after starting Eclipse.

While debugging, I noticed that GitProjectsImportPage#setProjectsList was called twice during the wizard. See https://git.eclipse.org/r/19477 for the fix for this, which seems to fix the issue with the project list being empty.

Patch-set: 1
Label: Verified=0