[stagingView] Fix flickering when switching between repos
When switching from a repo that is in the normal state to one that has
conflicts, a warning for "invalid committer" was briefly shown and then
hidden again, causing some flickering.

The reason was that the text fields are set to "" in case of conflicts,
which triggered the warning, only to be removed again later. So make
sure that setting to "" is done at the end and also don't cause any
"statusUpdated" events when the fields are not enabled.

Change-Id: I02dc7c9585d004496c39f635bc52d440fdf2cde2
Signed-off-by: Robin Stocker <robin@nibor.org>