Fix ContainerTreeIterator and RepositoryMapping for submodule repos
For submodule repositories, the first repository that is found that
contains the a path in its worktree may be the parent repository. So use
the one with the best match in RepositoryMapping.

In ContainerTreeIterator, just use the passed Repository for calculating
the prefix instead of searching for the mapping again.

Note that JGit change I0c00953f73f9316a66d0fc10eab52d8779c88f00 is also
necessary for the decorations to work correctly, see bug 423644.

Bug: 392378
Change-Id: Ib67a4593c49b29ed6d0c8357493f227707f7ad9e
Signed-off-by: Robin Stocker <robin@nibor.org>