Update patch set 2
Patch Set 2:

Ah yes, I was puzzled by a case where I have two entries for remote.origin.fetch in the same config file and with "single value" I see only one of them. But you are right, the command line's --get and --get-all behave the same like our "single value" button.

But on the other hand then we do not have a possibility to show the really "effective" config, i.d. the one which is valid. AFAIU there are two types of config parameters: either more than one value is allowed or not. remote.origin.fetch is an example of the first case, rebase.autostash of the latter case. For the latter case, the user cannot tell from the line [false][true] which of the two values is the valid one.

But I see that we do not have metadata for each of the config parameters to decide what to display as "effective". So I think we should leave it as it is.

Patch-set: 2