Use scrolling instead of soft-wrap in commit message text field
Wrapping is not desired for text like stack traces or code snippets.
Also, if hard-wrap is turned off, one may actually want to see the
message as it will be in the commit (most viewers don't soft-wrap the
commit message).

To only display scroll bars when necessary, also use
setAlwaysShowScrollBars(false).

Change-Id: I13175041866897960d434688f751fa20b18efa30
Signed-off-by: Robin Stocker <robin@nibor.org>