Update patch set 1
Patch Set 1:

Hello Stephan. With this patch, I see a strange behavior: When the "Show Find Toolbar" is enabled (so incremental loading disabled) the number of commits goes over the maximum set in the preferences. It seems it does a few more updates after detecting it's incomplete. I set my maximum to 20 commits to make it more apparent. This does not happen in my updated my patch but I'm not sure why since the approach is similar.

https://git.eclipse.org/r/#/c/20640/

Patch-set: 1