Update patch set 2
Patch Set 2:

This change is based on patch set 1 and tries to improve handling multiple repositories.

When there are multiple repositories that need a clean working copy before a successful pull can be performed, the dialog to stash/reset/etc. for the repositories are now opened one after another, instead of all at the same time.

The pull result dialog now waits for all pulls to finish (including those that were retried with a cleaned working copy), to present the results for all repositories in one dialog.

Patch-set: 2