Avoid cleaning if selection is empty.
The current implementation runs the git clean command even if no file or
directory is selected in the clean wizard. That may take quite some time
and may lead to cleaning the repository (if I understood the underlying
jgit command right), although the user expected nothing to be cleaned.

If there is a useful workflow for running the clean command without any
selection of to be cleaned paths, then this should be implemented by an
additional top level radio button "Clean all resources" or the like.

Change-Id: I719c682df62499fd7d02658fd87970ba7f006d51
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>