Update patch set 1
Patch Set 1:

I've moved the commit sorting method to the new class org.eclipse.egit.core.CommitUtil.

As for the commit message formatting, I have switched to the way the RebaseCommand does it. This unfortunately has the negative side effect that the complete squashing procedure cannot be canceled by simply clicking the "Cancel" button in the commit message dialog. This is because interactive rebase is starting right away, prompting for commit messages if it sees the need along the way.

Patch-set: 1