Update patch set 2
Patch Set 2:

Hi Robin and Matthias,

unfortunately this change worsens the situation on Windows for the commit message view. The diff view seems to work fine with the change.

The permanently shown, but inactive vertical scrollbar is gone, as intended by the fix. Instead there is now a a permanently active horizontal scrollbar. In addition to that, moving the mouse around in the commit message view causes both a horizontal and a vertical scrollbar to appear and flicker annoyingly.

As the original problem of showing an inactive scrollbar is only a minor one, I propose to change the code in CommitMessageViewer to

if (!SystemReader.getInstance().isWindows())
	t.setAlwaysShowScrollBars(false);

Patch-set: 2
Reviewer: Gerrit User 2957 <2957@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0