Update patch set 1
Patch Set 1: Code-Review-1

the "Push Branch..." wizard shouldn't display the warning "The existing upstream configuration for the branch will be overwritten." if the configuration options shown in the wizard match the persisted branch configuration. At the moment this warning is always shown if some configuration for this branch is persisted but I think this warning should only be displayed if the user changed any of the persisted configuration settings.

Patch-set: 1
Label: Code-Review=-1