Use a workspace-aware merging strategy when working from EGit
This introduces an Eclipse-specific sub-class of the
RecursiveMerger that will make use of the mergers provided
by ModelProviders instead of ignoring them to always use
EGit's custom merging algorithms.

Any file which does not have a specific merger (or which
merger cannot be found, such as for files outside of the
workspace) is still handled as it would by the
RecursiveMerger.

Signed-off-by: Laurent Goubet <laurent.goubet@obeo.fr>
Signed-off-by: Axel Richard <axel.richard@obeo.fr>
Also-by: Axel Richard <axel.richard@obeo.fr>
Change-Id: I1b27061ab215ebe600baef56f4e19a038c4b31a8