Update patch set 10
Patch Set 10: Code-Review-1

(1 comment)

The suggestion is only made if the new branch is based on a reference. i.e. if in CreateBranchPage myBaseCommit is null and myBaseRef != null (l. 270 in CreateBranchPage). When I create a branch on a commit in the history view where no branch or remote branch points to, therefore the suggestion is not made.

Is this the intended behavior?

Patch-set: 10
Reviewer: Gerrit User 6 <6@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1