Allow canceling the commit operation after save-during-commit
When a commit is performed, the user is asked to save all dirty editors.
Even if she does save some or all of them, the commit continues. The
changes that are saved, are not staged and thus not part of the commit.

If the user actually saved a file, we now ask her whether she wants to
cancel committing, to be able to stage the freshly saved files.

This only applies to commits via the staging view. The commit dialog
already allows the user to stage saved files without needing zo cancel.

Change-Id: Iec972afd5494019717d15a1d828e8569f5e40188
Signed-off-by: Konrad Kügler <swamblumat-eclipsebugs@yahoo.de>