Remove delay in local repositories filter
When trying to filter the directory with the Android open source
repositories (around 1400 repos) there is a delay of several seconds
after each update of the filter mask.

The performance issue is caused by getting the checked items during
filter evaluation of each item completely again and therefore has O(n^2)
runtime. This change caches the checked items after each change of a
checkmark.

With the change there is no notable delay when filtering the same list
of 1400 repositories.

Change-Id: Iab65452ae62551c3f8e7a2ee594a5a238b802144
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>