Revert "Sort refs in the 'Switch To' menu in natural order"
This reverts commit d6a9e12bc2535a4ed335ad426f5cebfc13e67cb8.

That change introduced a new bug. It leads to duplicated branches in the
SwitchTo menu. Unfortunately I cannot debug into that issue, as my
Ubuntu graphical desktop crashes when trying to do that.

Of course I'm happy if someone can instead provide a change to fix the
duplicated branch names issue and have the natural order.

Change-Id: Idc9d0fde481f243c00202aa5ada4a63ad709cf2c
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>