Update patch set 6
Patch Set 6:

I just noticed I forget the @Test on the test method. I also added assertions that show, that comparing for equality at the beginning is not enough. When the compared strings only differ by leading zeros in the number parts, the strings are not equals, but the comparison should return 0.

Patch-set: 6
Label: Verified=0