Update patch set 6
Patch Set 6:

Hm, thinking a bit more about it, two strings differing only by  zeros should not return 0. Because if it does return 0 and you have a branches named "b1" and "b01" only one of them would be shown in the switch-to menu.

Instead, when both string are equal disregarding zeros, one has to redo-the comparisons left-to-right and compare the number-parts as strings.

Since this is expensive, comparing with equalsIgnoreCase at the start would improve the (common) case that the two strings are actually the same.

What do you think?

Patch-set: 6