Update patch set 6
Patch Set 6:

We cannot just use a case ignoring compare here. TreeMaps's primary purpose is to eliminate duplicates which means it must not consider two different strings equals, ever. As it happened, it also sorted the items. Got two-in-one.

One way is to use a hashmap for eliminating duplicates and then sort.

Patch-set: 6