Update patch set 6
Patch Set 6: Code-Review-1

Robin Rosenberg is spot on.

This comparator "imposes orderings that are inconsistent with equals" (to use the terms in Comparator's documentation). The simplest case is "a" and "A", where equals is false, but they should be sorted the same.

To see the difference, run this in CommonUtils:

	public static void main(String[] args) {
		TreeMap<String, String> map = new TreeMap<String, String>(
				STRING_ASCENDING_COMPARATOR);
		map.put("a", "1");
		map.put("A", "2");
		map.put("B", "3");
		System.out.println(map);

		List<String> list = Arrays.asList("a", "A", "B");
		Collections.sort(list, STRING_ASCENDING_COMPARATOR);
		System.out.println(list);
	}

It returns the following:

  {a=2, B=3}
  [A, B, a]

So the fix is to change SwitchToMenu to use a normal HashSet for eliminating duplicates, and use Collections.sort for sorting the elements.

Patch-set: 6
Label: Code-Review=-1