[historyView] Prevent NPE for non-workspace comparisons
Using the actions "Compare with Version in Ancestor" and "Compare with
Workspace" in the history view for files resulted in a
NullPointerException, if the selected file was not in an imported
Eclipse project. In this case a null RepositoryMapping
was used to find the repo-relative path of the file to compare,
resulting in a NPE.

If no RepositoryMapping is available, use the repository we get as an
argument anyway, to get the repo-relative path.

Bug: 424585
Change-Id: I7b58b15f879b4f5d2b704faec87f05864a677199
Signed-off-by: Konrad Kügler <swamblumat-eclipsebugs@yahoo.de>