Update patch set 2
Patch Set 2:

I've added code locally to open the rebase view and set amending in the staging view. I *think* the code is correct, but I'm running into all sorts of problems here:

- When aborting the rebase, I get multiple IllegalStateExceptions.
- Setting the staging view to amending works, but it shows the wrong commit message. Clicking the "amend" button off and on again shows the correct message.

I'm not sure how to proceed here. Would it be okay to commit the changes into Gerrit so that you can have a look?

Patch-set: 2
Label: Verified=0