Show Annotations: Fix incorrect annotations when having local changes
How to reproduce:

1. Add a few lines in the middle of a tracked file
2. Save the file
3. Execute "Show Annotations"

Expected: The annotations for the lines below the added ones should be
correct.
Actual: They are at the wrong location (too far down)

The problem is that BlameCommand also reads the local file content when
startCommit is not set. But in Eclipse, local changes are already
handled by the Quick Diff provider. So before, they were accounted for
twice. By only blaming HEAD, the Quick Diff logic can work correctly.

Bug: 388543
Change-Id: I62e677d4308672ead349b6806d3d74e9723cb05c
Signed-off-by: Robin Stocker <robin@nibor.org>