Unregister RebaseInteractiveView from plan on view disposal
The view registers as a RebaseInteractivePlanChangeListener with the
plan, but it never unregistered. After closing the view on a running
interactive rebase and reopening it, it would notify the already
disposed old view, too. This resulted in a "Widget disposed" exception,
e.g. when using the buttons to move commits up or down in the rebase
view.

Also unregisters when another plan is shown.

Change-Id: I67b582e1add240bb4490e26c538f6e30c6c5e6e6
Signed-off-by: Konrad Kügler <swamblumat-eclipsebugs@yahoo.de>