Initially disable buttons in branch selection dialog
The enablement of the buttons in the branch selection dialog depends on
the highlighted initial branch. But if there is no initially highlighted
branch, then the delete and rename buttons should be disabled (as it
already happens for the okay button).

The buttons will automatically be enabled if there is some initial
branch highlighting, as that is evaluated afterwards (see existing
comment in line 327).

Change-Id: Ic22fbfbc7694f6f672f3426614fe0a0b2e68fb67
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>