Fix GitQuickDiffProvider creating multiple GitDocuments for resource
getReference didn't check whether a document was already created, and
just created a new one, overwriting the already existing one.

This resulted in the existing GitDocument never being disposed, which in
turn caused unnecessary work being done in GitDocument#onRefsChanged,
slowing down operations.

That GitQuickDiffProvider#getReference can be called multiple times can
be seen by adding a breakpoint and then doing this:

1. Show annotations in a tracked file -> first call
2. Edit the file
3. Commit the changes -> second call

The check for RepositoryProvider can be removed as it's done anyway in
GitDocument.create.

Bug: 413497
Change-Id: Iad7b39b85e365a7729316b6dbbbe99c25057147c
Signed-off-by: Robin Stocker <robin@nibor.org>