Don't look for .git in subdirectories of project
Connecting a project with a .git repository in one of its subdirectories
breaks the Team menu for the project and makes the subdirectory
disappear. When this happens with "auto share", it's very confusing.

Other places in the code could also not really handle this, e.g. the
share wizard displayed the repository location as ".git", when it really
should have shown "sub/dir/.git".

Bug: 421573
Change-Id: I4489cdd3526a30b1d2d575040792897a92da0c6d
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>