[stagingView] Don't clear restored commit message state in the UI
When viewing a repository in the Staging View for the first time after
starting Eclipse, the IndexDiff data for this repository is usually not
available yet. During StagingView.reload(), first the stored Staging
View state is restored. Then enableCommitWidgets(false) is called,
which disables everything, and worse, also clears the just restored
inputs, including the commit message.

When the IndexDiff for the repository becomes available, reload() will
be called again.
This change postpones restoring the view state to that time. This way
the restored commit (and amend etc.) state remains intact. With the
IndexDiff available, the inputs will not be cleared after restoring them
(unless there are conflicts to be resolved).

This change prepares for commit message restore after restarting
Eclipse.

Change-Id: I4fb24196bf07d4af9fa18b2a8390793e2057f6eb
Signed-off-by: Konrad Kügler <swamblumat-eclipsebugs@yahoo.de>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>