Update patch set 3
Patch Set 3:

if the other guy doesn't know you are working on his change this can lead to surprising race conditions and only the last patchset of a change can be submitted. Also I find it irritating if we end up with a change where the original author isn't listed as the author anymore. 

I sometimes take over a change when only small tweaks are open to finish the change and I can then submit it with some minor improvements or if I know the author is currently swamped and the solution to the remaining issues is pretty clear.

Patch-set: 3