Update patch set 4
Patch Set 4:

Not sure, other Eclipse projects do it differently I think. Our preferred format is documented in the contributor guide:

https://wiki.eclipse.org/EGit/Contributor_Guide#Commit_message_guidelines

But the general format of the commit message footer is generally agreed to AFAIK. I think there was a bug report/comment somewhere about commit message validation, but not sure. In any case, this is low prio IMO :).

Patch-set: 4