Update patch set 2
Patch Set 2:

> The choice should be disabled since "this" version does not exist.

I was also conflicted about this, but I'm not sure users really care about the distinction.

I'll have a look if the label can be changed in case of a deleted file.

> Alt. Have both "open this version" and "open previous version".

Hm, may this also be useful in general (for changed files)? Not sure if it's worth to increase the size of the menu for that.

Patch-set: 2
Label: Verified=0