[stagingView] Store the commit message on Eclipse shutdown
When closing Eclipse or only the staging view, the commit message (and
amend state, author, etc.) are stored now. Previously it was stored only
when switching the repository displayed in the staging view.

The restore part already exists and together with this change will now
prevent lost commit messages due to Eclipse restarts.

Bug: 402194
Change-Id: Ied7a3a65177a7fec5e728ac02e533d3e06e8c73f
Signed-off-by: Konrad Kügler <swamblumat-eclipsebugs@yahoo.de>