Update patch set 5
Patch Set 5: Code-Review-1

See my comment on patch set 4.

The follow filter is only set in GitHistoryPage's createFileWalker method if there is a filter set to file paths. Only then the rename tracker is reset to a path.

To make this change work if the filter is set to the whole repository , project or folder I think the follow filter also has to be set for these cases in order to fill the rename tracker.

Patch-set: 5
Label: Code-Review=-1