Update patch set 2
Patch Set 2:

I agree on the suggestion for the checkbox.
There seems to be an already existing problem that host parts are not supported in file URLs. That should be handled in a separate issue, because it requires more changes.

I think it is acceptable that the host is purged when changing the scheme.
This behavior is better IMO than not being able to delete the disabled host and being required to switch back the scheme.

Patch-set: 2
Reviewer: Gerrit User 966 <966@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0