Add support for cherry-picking multiple commits at once.
This modifies the Cherry Pick menu item in the History view such that
multiple selected commits may be cherry-picked at once, instead of
one-by-one.

If a conflict occurs during merging one of the commits, the whole
process will stop with an error message. The conflict then needs to be
resolved and committed. (This has not changed from the previous
behaviour.) After that, cherry-picking may be resumed manually with the
leftover commits.

Change-Id: I23d1339072c9557af530c9e3f8d6636cf24fbac1
Signed-off-by: Maik Schreiber <blizzy@blizzy.de>