Update patch set 4
Patch Set 4: Code-Review-1

(4 comments)

if "Push..." is clicked on a repository in the repositories view the commit message header is shown with the local commit ID, if it's clicked from a resource's team menu the commit message header is missing. We should it in both cases.

Looks like we could also use this simplified push dialog for "Push Commit..." in the history view.

Patch-set: 4
Label: Code-Review=-1