Add support for cherry-picking multiple commits at once.
This modifies the Cherry Pick menu item in the History view such that
multiple selected commits may be cherry-picked at once, instead of
one-by-one.

Cherry-picking now uses the mechanics of interactive rebase. If a
conflict occurs during merging one of the commits, it needs to be
resolved manually, and then the interactive rebase may be resumed
(just like with a regular interactive rebase.)

Cherry-picking a single commit uses the same mechanics just mentioned.

Change-Id: I23d1339072c9557af530c9e3f8d6636cf24fbac1
Signed-off-by: Maik Schreiber <blizzy@blizzy.de>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>