Update patch set 7
Patch Set 7:

I'm not sure the test failures are related to the suggested change. However, I tried to build locally and got many test failures, different from the one on Hudson.
Can you please re-trigger the build to check whether this is determistically failing, because of the suggest patch?

Patch-set: 7