Update patch set 14
Patch Set 14: Code-Review-2

(10 comments)

See inline comments.

Another comment: When the thing we are pushing is not a branch, the "Configure upstream" part must be hidden, because it's not possible to configure anything for a commit.

This is one more reason why it may be better to leave the old "ref push wizard" in place for pushing a commit, see inline comments. When we do that, we could instead show the "ref push wizard" on "Team > Push HEAD...", which should be easy with the dynamic menu.

Matthias, what do you think about this?

Patch-set: 14
Label: Code-Review=-2