Update patch set 17
Patch Set 17:

@Matthias: Didn't know that C Git is so strict :).

Ok then, as long as Mickael implements this in another change for 3.5, I'm ok with removing this for now. I just don't want users complaining that this functionality was removed without having a replacement.

(Would have been better to replace this in a separate change when the replacement is ready, but it's already done now.)

@Mickael: I would go with "make it unchecked and hidden when we are on a detached HEAD". Not sure if this is your second option. It should definitely stay for the branch case (upstream configuration is used for both push and pull).

Patch-set: 17