Update patch set 1
Patch Set 1: Code-Review+1

I think we could replace the implementation in SWTPlotRenderer with this in a follow-up. Although note that it treats the ellipsis as 3 characters, because in the worst case, it uses about the same amount of space as "lll".

But LGTM.

Patch-set: 1
Reviewer: Gerrit User 79 <79@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1