Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Oh, that's the same code duplicated in 3 places? Could you clean that up with this change?

I'd move the RepositoryActionHandler#canMerge method into a static method on MergeActionHandler that takes a Repository and a Shell, and then call it from the other 2 places.

Patch-set: 1
Label: Code-Review=-1