Update patch set 1
Patch Set 1:

There is one issue with this patch I should mention.  

When doing a "Compare with previous" of a file, the previous version used is the parent of the selected *commit*, which is not necessarily the commit which previously changed the file.  The comparison is correct, but the commit id is not the same one as the one shown in the file history view.

I didn't know of a simple way to find the previous commit when dealing with a single file.

Patch-set: 1
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0