Update patch set 1
Patch Set 1: Code-Review-1

I'd prefer to have a short (updated) description in this file so that the essentials can also be read offline. We can drop the description how to create patch files and instead give a short description how to push to Gerrit.

We could also consider to convert this file to markdown.

Patch-set: 1
Label: Code-Review=-1