Extract convenience methods in a separate class from the tests
This extracts a base abstract class for tests aimed at the
logical model support, and extract the Sample logical model
out of GitSubscriberContextTest to make it available for
other test classes.

The sample logical model provider is now properly
contributed through the extension point so that it will be
found and used by the platform during our tests the same way
it would be during regular use by an end-user.

Change-Id: Idd602bc7af278bdc47f38801388f533754a46819
Signed-off-by: Laurent Goubet <laurent.goubet@obeo.fr>
Signed-off-by: Axel Richard <axel.richard@obeo.fr>
Also-by: Axel Richard <axel.richard@obeo.fr>