Update patch set 2
Patch Set 2:

Some of those are implemented as actions currently (because they're also in the toolbar), which makes it more difficult. Last time I tried to convert them to the non-deprecated way I hit some platform limitations, see bug 414612.

But maybe we can just leave the toolbar items be actions and use org.eclipse.ui.menus for the menu. Will look into it. (But that should go in a separate change.)

Regarding "Add to Index" and "Remove from Index", would you expect them to operate on the whole repository or only on the connected projects that are in the working set? I personally would be a bit scared to use them on a whole working set, maybe they shouldn't be available.

"Ignore", on the working set? I don't think that's needed.

Patch-set: 2
Label: Verified=0