Update patch set 1
Patch Set 1: Code-Review-2

This interacts badly with the automatic selection of files if you show the history for a specific file. E.g. when the history for a specific file is shown, selecting a commit won't show the message, but go directly to the diff instead.

Another problem: The scrolling is not accurate, as the first line I see the last line of the commit message here, e.g. "Signed-off-by: ...". It seems to depend on the height of the bottom sash part of the history view.

Patch-set: 1
Reviewer: Gerrit User 79 <79@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-2