Make sure "Create Branch" dialog shows all contents
I don't know why, but the setMessage(null) call caused the dialog size
to be smaller than its contents. Change it so that the initial message
stays, which looks nicer. Also, don't show an initial empty name as an
error, which was the case when creating a branch for a commit from the
history view.

In case of creating a new branch from a commit and then choosing another
source in the dialog, resize the dialog to fit its contents.

Bug: 440097
Change-Id: I10d577450475a2bc7ce2d4b09b4b46d737369d24
Signed-off-by: Robin Stocker <robin@nibor.org>