Update patch set 2
Patch Set 2:

Yep, to make FB happy. In theory there could be code added to the try block after the lock(), which would make this warning real. 
Sorry for second patch set without changes, which seems to be a side effect of pushing another change to Gerrit.

Patch-set: 2
Label: Verified=0