Update patch set 1
Patch Set 1:

You'll have to learn not to mix different concerns/fixes in the same branch..

The simplest way is to create a new branch (based on master) and cherry-pick this commit there, then amend it and push it.

You can do the same for the second commit. When you push that, the review will be updated and you will see that the Dependencies will have changed (because you based the commit on master instead of this one).

Patch-set: 1
Label: Verified=0