[historyView] Use separate diff viewer instead of embedding into message
The diff viewer is the same as in the "Diff" tab of the commit viewer.
It still shares its scroll bar with the message.

The advantage is that when a file is selected, only the diff has to be
updated, not the whole message. We can also make the commit message
collapsible or extract other parts from the comment (such as the list of
branches/tags). Also, the hyperlink detectors now only run on the commit
message, not on the diff.

Bug: 440125
Change-Id: I6dc79f75eb0a63dde64cdf33d8e66fc594c327ab
Signed-off-by: Robin Stocker <robin@nibor.org>