[stagingView] Fix amend message being counted as "user-entered"
When clicking amend and then switching to another branch, the amend
state and commit message would stay the same.

This change fixes it so that after switching to another branch, amend
and commit message are reset. I think this is less surprising than
keeping amend but updating the message.

The other hunk in this change is so that we still preserve the amend
state when switching to another repository and back.

In case the user has modified the message after pressing amend, it is
discarded when switching to another branch.

Bug: 441607
Change-Id: I7af1bccde43ccd2048dae6643040170ee4b58c51
Signed-off-by: Robin Stocker <robin@nibor.org>