Fix repeating message 'Building workspace' in status bar
In Ibdc454f5ffb290831b81c5e775da35434811454c, many jobs were converted
from normal Job to WorkspaceJob.

In case of RepositoryChangeScanner which is run every few seconds, this
causes the message "Building workspace" to flash in the status bar. The 
reason for this is that a WorkspaceJob triggers an auto-build, even when 
no resources were changed.

As RepositoryChangeScanner does not change any resources itself, it
doesn't need to be a WorkspaceJob (which is for batching resource changed events). So change it back to a normal Job.

Bug: 439325
Change-Id: Ic555b868fbda952e7f88a1b851e63af2bb1d6545
Signed-off-by: Szymon Ptaszkiewicz <szymon.ptaszkiewicz@pl.ibm.com>