Update patch set 1
Patch Set 1:

Thanks for your contribution!

Could you please rebase both changes on the current master? As this is a new feature this is where they should be applied.

This change looks good and also fixes the case when you have selected commits with multiple actions and deselect commits so that you end up with a selection that has only one kind of action. In that case all buttons remained disabled. With this fix it correctly enables the button for the single action.

Patch-set: 1
Label: Verified=0