Update patch set 3
Patch Set 3:

It now removes a useful feature, which is this:

* Open History view with "Show all changes of selected resource and its children"
* Show the history for a file
* Select a commit

Expected: The commit message is shown and below it the diff for the file whose history is being shown.
Actual with this patch: Only the commit message is shown. For the diff, one has to click on the file first.

The expected behavior is useful when the user quickly wants to go through the commits and the diff for that file.

The change would be nice if it only revealed the diff when the *user* clicks on the file, but preserves the above feature.

Patch-set: 3