Update patch set 1
Patch Set 1: Code-Review-1

(4 comments)

That's pretty handy.

What I'm a bit worried about is that it will only find the visible branches/tags, so the user must be aware that they have to use "Show All Branches and Tags". But I think as all the search options work this way, it's ok.

The next step for improving the history view would be to have a dropdown where one can select the branch(es) that the history view should start walking at, as opposed to just current/all.

Patch-set: 1
Label: Code-Review=-1