Update patch set 1
Patch Set 1: Code-Review-1

Normally we read top-down. Git also used top-down in the rebase editor. We're used to that. We're not used to tortoise.

The reason the history is shown in reverse is that you are usually interested in the newest changes.

Patch-set: 1
Reviewer: Gerrit User 8 <8@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1