Update patch set 1
Patch Set 1:

"Growing up" with EGit and the History view, both a colleague of mine and I would find it easier if the rebase plan was presented in the same order as the history.

Maybe we should have a preference in the view menu of the Rebase Interactive view to accomodate for both mental models.

Patch-set: 1
Reviewer: Gerrit User 2957 <2957@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0