Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

With that the user doesn't yet have a chance to set the commit message for a merge commit since there is no way to enter the custom message. If you add that to the UI the merge commit message should be initialised with the generated standard message.

Patch-set: 2
Label: Code-Review=-1