Update patch set 2
Patch Set 2:

(2 comments)

I use EGit's MergeOperation only as a standalone operation, not in a UI way. I agree if I add that to the UI the merge commit message should be initialized with the generated standard message, but this is not the case because I use a standalone operation.
I can create an enhancement in bugzilla if you want.

Patch-set: 2
Label: Verified=0