Update patch set 1
Patch Set 1:

This works but I think it will be surprising for the user: cherry-pick was called but in order to resolve conflicts the user has to follow the rebase conflict resolution steps. This deviates from the steps needed for resolving a cherry-pick conflict using native git so I think we should explain that to the user.

Patch-set: 1