Update patch set 1
Patch Set 1:

When starting to cherry-pick, there is now a dialog that asks for confirmation of the commits selected (Robin put that dialog in.) I'm now thinking about adding an additional text to that dialog, along the lines of:

"Note: EGit uses interactive rebase mechanics for cherry-picking, which is different from regular Git. If one of the selected commits fails to merge, you need to follow conflict resolve steps as if doing a regular interactive rebase."

What do you think about that?

Patch-set: 1
Label: Verified=0