Introduce context menu "import changed projects"
The context menu action "import changed projects" analyzes the files of
the commit and is importing the matching projects. This is a handy
feature for very large repositories with lots of projects where it is
impossible to import all projects. Import only the relevant projects
needed for the selected change/commit.

Change-Id: I55aac6258659d61088e9abdafe95f5a4a6a6fe63
Signed-off-by: Tobias Melcher <tobias.melcher@sap.com>
Signed-off-by: Felix Otto <felix.otto@sap.com>
Signed-off-by: thomas.schulze.zur.wiesch@sap.com