Introduce context menu "import changed projects"
The context menu action "import changed projects" analyzes the files of
the commit and is importing the matching projects. This is a handy
feature for very large repositories with lots of projects where it is
impossible to import all projects. Import only the relevant projects
needed for the selected change/commit.

Change-Id: I70a1371fae61b37789be26e930128d1c50749708
Signed-off-by: Tobias Melcher <tobias.melcher@sap.com>