Update patch set 1
Patch Set 1: Code-Review-1

(14 comments)

I think this is a command which will only be used by a few users (those using a huge repository with many projects), hence I would like to avoid to add it to frequently used menus which are already pretty long. Hence I suggest to only add it to the history view as this is the place where it makes most sense as the command is inspecting a single commit as far as I understood.

Maybe you could also add it as an option to the already existing "Import Projects" command on repository nodes in the repositories view ?

Patch-set: 1
Label: Code-Review=-1