New EGit project logo (SVG)
Created in Inkscape based on the Eclipse logo and adding the graph part
from the Git logo.

This is the source of `org.eclipse.egit/egit.png`. Export to 32x32 was
done from Inkscape using -8 for x0 and y0 and 1040 for x1 and y1, to add
some additional margin when scaling down to 32x32.

Bug: 441460
Change-Id: I1412422764c7132e82fc5a8cbfcd08041d0bb6d5
Signed-off-by: Robin Stocker <robin@nibor.org>