Fix ContainerTreeIterator to return on-disk size of a folder
A ContainerTreeIterator always returned 0 as length of a directory. This
was inconsistent with JGits FileTreeIterator which asked the filesystem
object about the length. At least on unix filesystems the file-system
size of the directory was returned which was different from 0. On
Windows 0 was returned by java.io.File.
Because of this inconsistency folders containing submodules often have
been marked as dirty by mistake. The index was filled with data coming
from a FileTreeIterator (length!=0) while the ContainerTreeIterator
returned length==0.

This has been fixed by forcing ContainerTreeIterator to ask the
filesystem for the length of a folder.

Bug: 391280
Change-Id: I4b75ee1e875ead1d0d2f772056aca918d37d3ffd