Let staging view use submodule repos
If the staging view is linked to the selected resource and if that
resource belongs to a submodule then staging view should also run in
the context of the submodule repo. This allows to work with submodules
in a similar manner than with native git. If I want to do a commit in a
submodule repo then select a resource in that submodule repo and you can
use Staging view to prepare the commit in the submodule repo. In native
git you would "cd" to a folder containing a file from the submodule to
switch context from the containing repo to the submodule repo. In Egit
you have to select a resource from the submodule to achieve the same.

RepositoryMapping is enhanced by one method which can return a submodule
repository. In future RepositoryMapping could be further enhanced to 
support git submodules in general.

Bug: 382093
Bug: 446344
Change-Id: I6cbd5d5f4d0f57e93ce40a9dfb54e7b7c67e2e3a