Let staging view use submodule repos
If the staging view is linked to the selected resource and if that
resource belongs to a submodule then staging view should also run in
the context of the submodule repo. This allows to work with submodules
in a similar manner than with native git. If I want to do a commit in a
submodule repo then select a resource in that submodule repo and you can
use Staging view to prepare the commit in the submodule repo. In native
git you would "cd" to a folder containing a file from the submodule to
switch context from the containing repo to the submodule repo. In EGit
you have to select a resource from the submodule to achieve the same.

RepositoryMapping is enhanced by one method which can return a submodule
repository. In future RepositoryMapping could be further enhanced to 
support git submodules in general.

Bug: 382093
Bug: 446344
Change-Id: Ie61c597c87a68da16a92cd52fea0bc5a4ff5ec28
Also-By: Stefan Lay <stefan.lay@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>