Update patch set 4
Patch Set 4: Code-Review-1

(5 comments)

I think we should cache symlinks (and maybe also submodule entries) in IndexDiffData since typically there are only a few (or none) and running another walk to find them could be avoided. In order to enable this in an efficient way we could add the following method to IndexDiff

 Set<String> getPathsWithIndexMode(FileMode mode)

which IndexDiffData could use to retrieve symlink and submodule entries from the DirCache. This method should respect IndexDiffFilters so that we don't spoil the incremental updates supported by IndexDiffData triggered by resourceChangedEvents.

Looks like some change events are missed in staging view when deleting symlinks.

When deleting a symlink referring to a directory the deletion isn't automatically staged which looks inconsistent compared to the deletion of a symlink referring to a file which is staged automatically.

Patch-set: 4
Label: Code-Review=-1