Force refreshing IndexDiffCache on any change on a .gitattribute file
After a .gitattributes file has changed we need to update the
IndexDiffCache to update the staging view and the decorators. For
example, the value of the ident attribute can change the result of the
isModified method on the WorkingTreeIterator.

This review is intented to be integrated after the 3 reviews regarding
git attributes on the JGit repository:
* Add basic support for attributes.
  I05f3ce1861fbf9896b1bcb7816ba78af35f3ad3d
  
* [342372]Adds the git attributes computation on the treewalk.
  I5e33257fd8c9895869a128bad3fd1e720409d361
  
* [342372] Ident attribute implementation
  Ia0ff2775223b028dfc54b6733f46eabb2520149c

Change-Id: I25aba8051aeec81b933e228412cab6221148e054
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>