Display git attributes in the git property page
Adds tables displaying the attributes of an IResource in the git
property page. Depending on the type of attribute for the current entry
the display may change:
* If there no attributes a "None." label is displayed
* If the attributes are identical for both checkout and checkin
operation on unique table is displayed.
* If there is any differences between the checkin and the checkout
attributes, two tables are displayed.

This review is intented to be integrated after the 3 reviews regarding
git attributes on the JGit repository:
* Add basic support for attributes.
  I05f3ce1861fbf9896b1bcb7816ba78af35f3ad3d
  
* [342372]Adds the git attributes computation on the treewalk.
  I5e33257fd8c9895869a128bad3fd1e720409d361
  
* [342372] Ident attribute implementation
  Ia0ff2775223b028dfc54b6733f46eabb2520149c

Change-Id: I20ef5868db27c464be1b641b0792abbddf031495
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>