Update patch set 1
Patch Set 1: Code-Review+1

I was able to successfully test this patch by cherry-picking it on top of 3.5.0, which is used by my test environment. 

BTW, for SSE-based editor the 'Show Annotations' action is available in the Team submenu of the ruler's context menu instead of having it directly in the ruler's context menu. Is it supposed to be like this?

Patch-set: 1
Reviewer: Gerrit User 621 <621@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1