Fix Create Eclipse Workspace Patch Crashes
Creating a patch in format "Eclipse Workspace" sometimes does not work.
The wizard simply doesn't do anything. 

The issue happens in two cases:
1. One of the files in the commit is NOT contained in any Eclipse
project.
2. One of the files in the commit is contained in MORE THAN ONE Eclipse
project.

There is no reason not to create a patch in the second case. In fact it
doesn't matter from which project the file is considered.


Bug: 450283
Change-Id: Ifc03b9d105122785713cad20291212ee75849a50
Signed-off-by: Reiner Hille-Doering <reiner.hille-doering@sap.com>