Update patch set 6
Patch Set 6:

> deletion of a symlink referring to a directory isn't auto staged, this feels inconsistent to the handling of file or symlink to file deletions

Ok, I raised a new bug for the problem, because it is not directly related with the symlink icon.
https://bugs.eclipse.org/bugs/show_bug.cgi?id=450525

> deleted symlink referring to a directory is displayed with a decorated file icon but it should use a decorated directory icon

Ok, but to know if the deleted element is a symlink referring to a directory, we need to retrieve the index  version of the file. Potentially, it is costly to do that in the stagingViewLabelProvider.
Do we really want to do that even so ?

Patch-set: 6
Reviewer: Gerrit User 1013 <1013@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0