Update patch set 2
Patch Set 2:

I wasn't suggesting that, only that if we would need a 3.5 for another reason, it would be good to have it included. But we could always cherry-pick it if that happens.

Are you planning on merging master into stable-3.6 some time before the 3.6 release? Or does everything for 3.6 need to target stable-3.6?

Patch-set: 2