Consider the preference regarding the logical model in EGit operations
This moves the property regarding the usage of the logical model from
egit.ui to egit.core so that it can be considered when setting the merge
strategy of the JGit commands within the EGit operations. Therefore,
getting the merge strategy is refactored out into a new OperationUtil,
which will be asked for the merge strategy to be used by the operations.
In OperationUtil, the preference regarding the logical model will be
considered before returning either the merge strategy, which enables the
logical model, or the default strategy (MergeStrategy.RECURSIVE).

Change-Id: I251e20d799f4806d3e1f920440622ef810f91a06
Signed-off-by: Philip Langer <planger@eclipsesource.com>