Fix CherryPickOperation in case of single commit cherry-picks
In case of cherry-picking on top of a merge commit there was a severe
bug leading to wrong results. This should be fixed by better support for
multi-commit cherry-pick in jgit. This commit introduces a fix which
solves the problem for single-commit cherry-picks (most of the people
only do that) and still supports to do multi-commit cherry-picks on
non-merge heads.

Change-Id: If5fbcc081cebfa01d9a202fa324c9996408e9062
Depends-On: jgit/jgit~master:I22b45a0a08eecb250ffd62dc487221f7f27def82
Bug: 447115