Use baseline instead of centerline in PlotRenderer
If the text extent height of a to be rendered plot line is odd, then the
SWTPlotRenderer cannot calculate the correct Y position for drawing the
label and draws the label with a 1 pixel offset. SWT text drawing uses
the baseline as Y coordinate. Due to the given centerline API in the
AbstractPlotRenderer the overall calculation of the baseline for SWT is
effectively (height / 2) * 2, thereby rounding all odd heights downward
to the next even number.

This change pushes the division by 2 from the caller into the
implementations of drawText.

In addition, enabling antialiasing has been moved from the
SWTPlotRenderer to the CommitGraphTable itself. Otherwise the cell label
heights are calculated different (1 pixel more height with antialiasing
enabled) in the middle of painting a row. This issue seems to be
specific to Windows.

Changes have been tested on Ubuntu14.10/GTK and Windows 7.

Bug: 450813
JGit-Dependency: I66f4e71873bb8e6f936fde573bbe4c35fe23a022
Change-Id: Iae782a8554bde6f812b0d8b374e70dc80cb46db6
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>