Update patch set 1
Patch Set 1: Code-Review-1

I think the intention to show the base version in the right hand pane was that the other two stages are shown already in the left hand pane. So it's kind of a poor man's 3 way compare.

If we do what you propose here we should offer 3 way compare like in the merge tool and show the same stage in the right pane like the merge tool does (currently it shows the other branch than your change. We should also show the shortened commit message in the title to make it easier to recognize which version is shown in which pane.

Alternatively we could simply correct the title of the right pane to "Common base version" or "Common ancestor".

Ideally we should have a 4 way compare which can show all 3 stages from the index and the workspace content in a 4th pane. p4merge [1] is a tool doing that in a nice way.

[1] http://www.perforce.com/product/components/perforce-visual-merge-and-diff-tools

Patch-set: 1
Label: Code-Review=-1