Update patch set 1
Patch Set 1:

> I think the intention to show the base version in the right hand pane was that the other two stages are shown already in the left hand pane. So it's kind of a poor man's 3 way compare.

I don't think there was any intention when this was chosen, apart from "CONFLICT" files are shown in unstaged changes, so compare against index like the other cases. It just so happens that the first index entry is the base stage.

Also, "two stages are shown already in the left hand pane" -> this is no longer the case after you started resolving conflicts.

I'm not sure. When someone wants the merge tool behavior, they can just use the merge tool instead, no? Or what would be the difference between this new compare and merge tool?

> Alternatively we could simply correct the title of the right pane to "Common base version" or "Common ancestor".

-1. I've heard from several people that the right pane "makes no sense", including the bug reporter. I myself sometimes just staged the conflict and then did compare to see the diff I actually wanted.

For modified files, comparing shows the differences to what is committed (or staged in case of partial staging). This is just as useful for conflicting files, for a "what's the effective result of this merge?".

So, in summary, I think the merge tool should be used for a nice real 3-way compare, and compare should be a simple "what would be the changes if I commit the file", as in other cases.

(Btw, I'll be on vacation the next few weeks, so won't be able to do any changes, sorry :).)

Patch-set: 1
Reviewer: Gerrit User 79 <79@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0